% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_functions.r
\name{wolfram_rule}
\alias{wolfram_rule}
\title{Create the rule for a specific Wolfram number}
\usage{
wolfram_rule(rule)
}
\arguments{
\item{rule}{the Wolfram rule}
}
\value{
a vector with 8 elements defining the responses to:
        (111), (110), (101), (100), (011), (010), (001), (000)
        on the previous row
}
\description{
Create the rule for a specific Wolfram number
}
\examples{
# get the definition of rule 30
wolfram_rule(30)

}
