% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_functions.r
\name{wolfram_rule_def}
\alias{wolfram_rule_def}
\title{Plot the definition of a Wolfram rule}
\usage{
wolfram_rule_def(rule)
}
\arguments{
\item{rule}{integer, the Wolfram rule}
}
\value{
a ggplot object defining the rule
}
\description{
Plot the definition of a Wolfram rule
}
\examples{
wolfram_rule_def(30)

}
